[App/Ruby] Move Ruby integration out into an XPC service.
