+++ app/CPReflectionService/CPReflectionService.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+#import "CPReflectionServiceProtocol.h"
+
+@interface CPReflectionService : NSObject <CPReflectionServiceProtocol>
+@end
