+++ app/CPReflectionService/RBObject+CocoaPods.h
@@ -1,15 +1,7 @@
 #import <Foundation/Foundation.h>
 #import <RubyCocoa/RBObject.h>
 
-extern NSString * _Nonnull const CPErrorDomain;
-extern NSString * _Nonnull const CPErrorRubyBacktrace;
-extern NSString * _Nonnull const CPErrorObjCBacktrace;
-
-typedef NS_ENUM(NSInteger, CPErrorDomainCode) {
-  CPInformativeError, // These are user errors
-  CPStandardError,    // These are runtime errors
-  CPNonRubyError      // These are errors on the Objective-C side of the bridge
-};
+NSError * _Nonnull CPErrorFromException(NSException * _Nonnull exception, NSString * _Nullable message);
 
 #pragma mark - Ruby integration
