+++ app/CPReflectionService/RBObject+CocoaPods.rb
@@ -1,9 +1,9 @@
 # This is required for Foundation classes to be known to this side of the bridge at all.
 require 'osx/objc/foundation'
 
-app_bundle = OSX::NSBundle.mainBundle
-bundle_path = File.join(app_bundle.resourcePath, 'bundle')
-incorrect_root = File.join(app_bundle.bundlePath, 'Contents/MacOS')
+service_bundle = OSX::NSBundle.mainBundle
+bundle_path = File.expand_path('../../Resources/bundle', service_bundle.bundlePath)
+incorrect_root = File.join(service_bundle.bundlePath, 'Contents/MacOS')
 
 # Fix all load paths to point to the bundled Ruby.
 $LOAD_PATH.map! do |path|
