+++ app/CocoaPods/CPAppDelegate.h
@@ -1,4 +1,5 @@
 #import <Cocoa/Cocoa.h>
 
 @interface CPAppDelegate : NSObject <NSApplicationDelegate>
+@property (readonly) NSXPCConnection *reflectionService;
 @end
